% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{all_dmls}
\alias{all_dmls}
\title{The complete Distance Metric Learning algorithms list.}
\format{An object of class \code{character} of length 20.}
\usage{
all_dmls
}
\description{
The complete Distance Metric Learning algorithms list.
}
\keyword{datasets}
