% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.numeric_or_null}
\alias{as.numeric_or_null}
\title{Converts object to numeric, preserving the NULL value.}
\usage{
as.numeric_or_null(x)
}
\arguments{
\item{x}{Object to cast.}
}
\value{
x casted, or NULL.
}
\description{
Converts object to numeric, preserving the NULL value.
}
