% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.type_or_null}
\alias{as.type_or_null}
\title{Converts object to certain type, preserving the NULL value.}
\usage{
as.type_or_null(x, as.type)
}
\arguments{
\item{x}{Object to cast.}

\item{as.type}{A casting function for x}
}
\value{
x casted, or NULL.
}
\description{
Converts object to certain type, preserving the NULL value.
}
