% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{methods}
\name{as.character.idType}
\alias{as.character.idType}
\title{Convert \code{idType} object to \code{character}}
\usage{
\method{as.character}{idType}(x, ...)
}
\arguments{
\item{x}{\code{idType} object.}

\item{...}{Further arguments to be passed.}
}
\description{
Function to convert \code{idType} object to \code{character}.
}
\keyword{manip}
