% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{methods}
\name{as.character.reactIdType}
\alias{as.character.reactIdType}
\title{Convert \code{reactIdType} object to \code{character}}
\usage{
\method{as.character}{reactIdType}(x, ...)
}
\arguments{
\item{x}{\code{reactIdType} object.}

\item{...}{Further arguments to be passed.}
}
\description{
Function to convert \code{reactIdType} object to \code{character}.
}
\keyword{manip}
