% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_wrappers.R
\name{AsDendrogram}
\alias{AsDendrogram}
\title{\code{RDML$AsDendrogram()} wrapper}
\usage{
AsDendrogram(obj, ...)
}
\arguments{
\item{obj}{\code{RDML} object.}

\item{...}{\code{AsDendrogram} params.}
}
\description{
Read more at \link{RDML.AsDendrogram}
}

