% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_wrappers.R
\name{AsTable}
\alias{AsTable}
\title{\code{RDML$AsTable()} wrapper}
\usage{
AsTable(obj, ...)
}
\arguments{
\item{obj}{\code{RDML} object.}

\item{...}{\code{AsTable} params.}
}
\description{
Read more at \link{RDML.AsTable}
}
