% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{base_dmls}
\alias{base_dmls}
\title{Base Distance Metric Learning algorithms.}
\format{An object of class \code{character} of length 3.}
\usage{
base_dmls
}
\description{
Base Distance Metric Learning algorithms.
}
\keyword{datasets}
