% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{cdnaSynthesisMethodType}
\alias{cdnaSynthesisMethodType}
\title{cdnaSynthesisMethodType R6 class.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Description of the cDNA synthesis method. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{cdnaSynthesisMethodType$new(enzyme = NULL,
  primingMethod = NULL, dnaseTreatment = NULL, thermalCyclingConditions =
  NULL)}

  @section Fields: \describe{
  \item{\code{enzyme}}{\link[checkmate]{checkString}. Name of the enzyme used for
  reverse transcription.}
  \item{\code{primingMethod}}{\link{primingMethodType}.}
  \item{\code{dnaseTreatment}}{\link[checkmate]{checkFlag} if \code{TRUE}RNA was
  DNAse treated prior cDNA synthesis.}
  \item{\code{thermalCyclingConditions}}{\link{idReferencesType}.}
  }
}

\section{Super class}{
\code{\link[RDML:rdmlBaseType]{RDML::rdmlBaseType}} -> \code{cdnaSynthesisMethodType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cdnaSynthesisMethodType-new}{\code{cdnaSynthesisMethodType$new()}}
\item \href{#method-cdnaSynthesisMethodType-clone}{\code{cdnaSynthesisMethodType$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id=".asXMLnodes"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-.asXMLnodes'><code>RDML::rdmlBaseType$.asXMLnodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="copy"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-copy'><code>RDML::rdmlBaseType$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="print"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-print'><code>RDML::rdmlBaseType$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cdnaSynthesisMethodType-new"></a>}}
\if{latex}{\out{\hypertarget{method-cdnaSynthesisMethodType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cdnaSynthesisMethodType$new(
  enzyme = NULL,
  primingMethod = NULL,
  dnaseTreatment = NULL,
  thermalCyclingConditions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cdnaSynthesisMethodType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cdnaSynthesisMethodType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cdnaSynthesisMethodType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
