% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{classifiers}
\alias{classifiers}
\title{Distance based classifiers.}
\format{An object of class \code{character} of length 3.}
\usage{
classifiers
}
\description{
Distance based classifiers.
}
\keyword{datasets}
