% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.Clone.R
\docType{methods}
\name{Clone}
\alias{Clone}
\alias{RDML.Clone}
\title{Clones \code{RDML} objects}
\arguments{
\item{to.merge}{\code{RDML} object that should be merged into current object.}
}
\description{
Clones \code{RDML} object for propertly usage.
}
\examples{
\dontrun{
PATH <- path.package("RDML")
filename <- paste(PATH, "/extdata/", "lc96_bACTXY.rdml", sep ="")
lc96 <- RDML$new(filename)
filename <- paste(PATH, "/extdata/", "stepone_std.rdml", sep ="")
stepone <- RDML$new(filename)
lc96$Merge(stepone)
lc96$AsTable()
}
}

