% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{commercialAssayType}
\alias{commercialAssayType}
\title{commercialAssayType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
commercialAssayType$new(company, orderNumber)
}
\arguments{
\item{company}{\link[assertthat]{is.string}.}

\item{orderNumber}{\link[assertthat]{is.string}.}
}
\description{
For some commercial assays, the primer sequences may be unknown. This element
allows to describe commercial assays.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

