% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{dataCollectionSoftwareType}
\alias{dataCollectionSoftwareType}
\title{dataCollectionSoftwareType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
dataCollectionSoftwareType$new(name, version)
}
\arguments{
\item{name}{\link[assertthat]{is.string}.}

\item{version}{\link[assertthat]{is.string}.}
}
\description{
Software name and version used to collect and analyze the data.\cr Inherits:
\link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

