% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{dimensionality_reduction_dmls}
\alias{dimensionality_reduction_dmls}
\title{The Distance Metric Learning algorithms oriented to dimensionality reduction.}
\format{An object of class \code{character} of length 3.}
\usage{
dimensionality_reduction_dmls
}
\description{
The Distance Metric Learning algorithms oriented to dimensionality reduction.
}
\keyword{datasets}
