% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_clf.R
\docType{data}
\name{distance_clf}
\alias{distance_clf}
\title{The distance classifiers global list.}
\format{An object of class \code{list} of length 3.}
\usage{
distance_clf
}
\description{
This list contains all the distance classifier constructors.
}
\keyword{datasets}
