% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\docType{data}
\name{dml}
\alias{dml}
\title{The DML global list.}
\format{An object of class \code{list} of length 20.}
\usage{
dml
}
\description{
This list contains all the DML Algorithm constructors.
}
\keyword{datasets}
