% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{DML_eig}
\alias{DML_eig}
\title{Distance Metric Learning with Eigenvalue Optimization (DML-eig).}
\usage{
DML_eig(mu = 1e-04, tol = 1e-05, eps = 1e-10, max_it = 25)
}
\arguments{
\item{mu}{Smoothing parameter. Float.}

\item{tol}{Tolerance stop criterion (difference between two point iterations at gradient descent). Float.}

\item{eps}{Precision stop criterion (norm of gradient at gradient descent). Float.}

\item{max_it}{Number of iterations at gradient descent. Integer.}
}
\value{
The LSI transformer, structured as a named list.
}
\description{
A DML Algorithm that learns a metric that minimizes the minimum distance between different-class points
constrained to the sum of distances at same-class points be non higher than a constant.
}
\references{
Yiming Ying and Peng Li. “Distance metric learning with eigenvalue optimization”. In: Journal of
            Machine Learning Research 13.Jan (2012), pages 1-26.
}
