% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plot.R
\docType{data}
\name{dml_plotter}
\alias{dml_plotter}
\title{The tune algorithms global list.}
\format{An object of class \code{list} of length 3.}
\usage{
dml_plotter
}
\description{
This list contains all the tune functions.
}
\keyword{datasets}
