% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\docType{data}
\name{dml_tune}
\alias{dml_tune}
\title{The tune algorithms global list.}
\format{An object of class \code{list} of length 2.}
\usage{
dml_tune
}
\description{
This list contains all the tune functions.
}
\keyword{datasets}
