% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{DMLMJ}
\alias{DMLMJ}
\title{Distance Metric Learning through the Maximization of the Jeffrey divergence (DMLMJ).}
\usage{
DMLMJ(num_dims = NULL, n_neighbors = 3, alpha = 0.001,
  reg_tol = 1e-10)
}
\arguments{
\item{num_dims}{Dimension desired for the transformed data. If NULL, dimension will be the number of features.}

\item{n_neighbors}{Number of neighbors to consider in the computation of the difference spaces.}

\item{alpha}{Regularization parameter for inverse matrix computation.}

\item{reg_tol}{Tolerance threshold for applying regularization. The tolerance is compared with the matrix determinant.}
}
\value{
The DMLMJ transformer, structured as a named list.
}
\description{
A DML Algorithm that obtains a transformer that maximizes the Jeffrey divergence between
the distribution of differences of same-class neighbors and the distribution of differences between
different-class neighbors.
}
\references{
Bac Nguyen, Carlos Morell and Bernard De Baets. “Supervised distance metric learning through
            maximization of the Jeffrey divergence”. In: Pattern Recognition 64 (2017), pages 215-225.
}
