% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{documentationType}
\alias{documentationType}
\title{documentationType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
documentationType$new(id, text = NULL)
}
\arguments{
\item{id}{\link{idType}. Identificator.}

\item{text}{\link[assertthat]{is.string}. Text.}
}
\description{
These elements should be used if the same description applies to many
samples, targets or experiments.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

