% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{Euclidean}
\alias{Euclidean}
\title{Euclidean distance.}
\usage{
Euclidean()
}
\value{
The Euclidean (identity) transformer, structured as a named list.
}
\description{
A basic transformer that represents the euclidean distance.
}
