% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{experimenterType}
\alias{experimenterType}
\title{experimenterType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
experimenterType$new(id, firstName, lastName, email = NULL, labName =
  NULL, labAddress = NULL)
}
\arguments{
\item{id}{\link{idType}. Identificator.}

\item{firstName}{\link[assertthat]{is.string}. First name.}

\item{lastName}{\link[assertthat]{is.string}. Last name.}

\item{email}{\link[assertthat]{is.string}. Email.}

\item{labName}{\link[assertthat]{is.string}. Lab name.}

\item{labAddress}{\link[assertthat]{is.string}. Lab address.}
}
\description{
Contact details of the experimenter.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

