% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.R
\docType{methods}
\name{[.GetFData}
\alias{[.GetFData}
\alias{[.RDML}
\title{Extract data points from \code{RDML} object}
\usage{
\method{[}{RDML}(x, i, j, dp.type = "adp")
}
\arguments{
\item{x}{\code{RDML} object.}

\item{i, j}{indices.}

\item{dp.type}{Type of fluorescence data (i.e. 'adp' for qPCR or 'mdp' for
melting).}
}
\description{
Extract data points from \code{RDML} object as.data.frame.
}
\keyword{manip}
