% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{generic_dmls}
\alias{generic_dmls}
\title{Generic Distance Metric Learning algorithms.}
\format{An object of class \code{character} of length 3.}
\usage{
generic_dmls
}
\description{
Generic Distance Metric Learning algorithms.
}
\keyword{datasets}
