% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_wrappers.R
\name{GetFData}
\alias{GetFData}
\title{\code{RDML$GetFData()} wrapper}
\usage{
GetFData(obj, ...)
}
\arguments{
\item{obj}{\code{RDML} object.}

\item{...}{\code{GetFData} params.}
}
\description{
Read more at \link{RDML.GetFData}
}
