% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{information_theory_based_dmls}
\alias{information_theory_based_dmls}
\title{The Distance Metric Learning algorithms based on information theory.}
\format{An object of class \code{character} of length 3.}
\usage{
information_theory_based_dmls
}
\description{
The Distance Metric Learning algorithms based on information theory.
}
\keyword{datasets}
