% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{KANMM}
\alias{KANMM}
\title{The kernelized version of ANMM.}
\usage{
KANMM(num_dims = NULL, n_friends = 3, n_enemies = 1,
  kernel = "linear", gamma = NULL, degree = 3, coef0 = 1,
  kernel_params = NULL)
}
\arguments{
\item{num_dims}{Dimension desired for the transformed data. Integer. If NULL, all features will be taken. Integer.}

\item{n_friends}{Number of nearest same-class neighbors to compute homogeneus neighborhood. Integer.}

\item{n_enemies}{Number of nearest different-class neighbors to compute heterogeneus neigborhood. Integer.}

\item{kernel}{Kernel to use. Allowed values are: "linear" | "poly" | "rbf" | "sigmoid" | "cosine" | "precomputed".}

\item{gamma}{Kernel coefficient for rbf, poly and sigmoid kernels. Ignored by other
kernels. Default value is 1/n_features. Float.}

\item{degree}{Degree for poly kernels. Ignored by other kernels. Integer.}

\item{coef0}{Independent term for poly and sigmoid kernels. Ignored by other kernels. Float.}

\item{kernel_params}{Parameters (keyword arguments) and values for kernel passed as
callable object. Ignored by other kernels.}
}
\value{
The KANMM transformer, structured as a named list.
}
\description{
The kernelized version of ANMM.
}
\references{
Fei Wang and Changshui Zhang. “Feature extraction by maximizing the average neighborhood
            margin”. In: Computer Vision and Pattern Recognition, 2007. CVPR’07. IEEE Conference on.
            IEEE. 2007, pages 1-8.
}
