% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{kernel_dmls}
\alias{kernel_dmls}
\title{Kernel Distance Metric Learning algorithms.}
\format{An object of class \code{character} of length 4.}
\usage{
kernel_dmls
}
\description{
Kernel Distance Metric Learning algorithms.
}
\keyword{datasets}
