% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_clf.R
\name{kNN}
\alias{kNN}
\title{k-Nearest Neighbors (kNN).}
\usage{
kNN(n_neighbors, dml_algorithm)
}
\arguments{
\item{n_neighbors}{Number of neighbors to consider in classification.}

\item{dml_algorithm}{The distance metric learning algorithm to use.}
}
\value{
The kNN classifier, structured as a named list.
}
\description{
The nearest neighbors classifier adapted to be used with distance metric learning algorithms.
}
