% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{knn_oriented_dmls}
\alias{knn_oriented_dmls}
\title{The Distance Metric Learning algorithms oriented to k-NN classifier.}
\format{An object of class \code{character} of length 2.}
\usage{
knn_oriented_dmls
}
\description{
The Distance Metric Learning algorithms oriented to k-NN classifier.
}
\keyword{datasets}
