% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{LDA}
\alias{LDA}
\title{Linear Discriminant Analysis (LDA).}
\usage{
LDA(num_dims = NULL, thres = NULL)
}
\arguments{
\item{num_dims}{Number of components (< n_classes - 1) for dimensionality reduction. If None, it will be taken as n_classes - 1. Ignored if thres is provided. Integer.}

\item{Fraction}{of variability to keep, from 0 to 1. Data dimension will be reduced until the lowest dimension that keeps 'thres' explained variance. Float.}
}
\value{
The LDA transformer, structured as a named list.
}
\description{
A distance metric learning algorithm for supervised dimensionality reduction, maximizing the ratio of variances between classes and within classes.
}
