% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{measureType}
\alias{measureType}
\title{measureType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
measureType$new(value)
}
\arguments{
\item{value}{\link[assertthat]{is.string}.}
}
\description{
Can take values:
\describe{
\item{real time}{}
\item{meltcurve}{}
}
\cr Inherits: \link{enumType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

