% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{Metric}
\alias{Metric}
\title{Basic learning from a metric.}
\usage{
Metric(metric)
}
\arguments{
\item{metric}{(d x d) matrix. A positive semidefinite matrix, to define a pseudodistance in euclidean d-dimensional space.}
}
\value{
@return The Metric transformer, structured as a named list.
}
\description{
A DML algorithm that defines a distance given a PSD metric matrix.
}
