% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/names.RDML.R
\name{names.RDML}
\alias{names,RDML-method}
\alias{names.RDML}
\title{The Names of the RDML object}
\usage{
\method{names}{RDML}(x, ...)
}
\arguments{
\item{x}{an object of class \code{RDML}.}

\item{...}{other arguments passed to \code{\link[base]{ls}} function.}
}
\description{
Just wraps \code{ls()} function to show RDML public fields
and methods. Useful for auto complete in R editors.
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>, Stefan Roediger
<stefan.roediger@b-tu.de>, Michal Burdukiewicz
<michalburdukiewicz@gmail.com>
}

