% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{ncm_oriented_dmls}
\alias{ncm_oriented_dmls}
\title{The Distance Metric Learning algorithms oriented to nearest centroids classifiers.}
\format{An object of class \code{character} of length 2.}
\usage{
ncm_oriented_dmls
}
\description{
The Distance Metric Learning algorithms oriented to nearest centroids classifiers.
}
\keyword{datasets}
