% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_clf.R
\name{NCMC_Classifier}
\alias{NCMC_Classifier}
\title{Nearest Class with Multiple Centroids classifier.}
\usage{
NCMC_Classifier(centroids_num = 3, ...)
}
\arguments{
\item{centroids_num}{If it is a list, it must have the same size as the number of classes. In this case, i-th item will be the number of
centroids to take in the i-th class. If it is an int, every class will have the same number of centroids.}

\item{...}{Additional arguments for Scikit-Learn K-Means.}
}
\value{
The NCMC classifier, structured as a named list.
}
\description{
A classifier that makes its predictions by choosing the class who has a centroid the nearest to the point.
For each class, an arbitrary number of centroids can be set. This centroids are calculated using k-Means
over each class sub-dataset.
}
