% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{new}
\alias{RDML.new}
\alias{new}
\title{Creates new instance of \code{RDML} class object}
\arguments{
\item{rdmlfile}{RDML file with fluorescence data.}

\item{name.pattern}{sample name pattern (see 'Details' at \link{RDML}).}
}
\description{
This function has been designed to import fluorescence data from instruments
that support RDML v1.1 format export.
}
\section{Warning}{
 Although the format RDML claimed as data exchange format,
  the specific implementation of the format at devices from different
  manufacturers differ significantly. Currently this function is checked
  against RDML data from devices: \emph{Bio-Rad CFX96}, \emph{Roche
  LightCycler 96} and \emph{Applied Biosystems StepOne}.
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>, Stefan Roediger
  <stefan.roediger@hs-lausitz.de>, Michal Burdukiewicz
  <michalburdukiewicz@gmail.com>
}

