% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{non_supervised_dmls}
\alias{non_supervised_dmls}
\title{Non supervised Distance Metric Learning algorithms.}
\format{An object of class \code{character} of length 1.}
\usage{
non_supervised_dmls
}
\description{
Non supervised Distance Metric Learning algorithms.
}
\keyword{datasets}
