% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{nucleotideType}
\alias{nucleotideType}
\title{nucleotideType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
nucleotideType$new(value)
}
\arguments{
\item{value}{\link[assertthat]{is.string}. Value.}
}
\description{
Can take values:
\describe{
\item{DNA}{}
\item{genomic-DNA}{}
\item{cDNA}{}
\item{RNA}{}
}
}
\details{
\cr Inherits: \link{enumType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

