% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numeric.as.integer}
\alias{numeric.as.integer}
\title{Converts every numeric item in the list to integer.}
\usage{
numeric.as.integer(x)
}
\arguments{
\item{x}{List.}
}
\value{
A list with every numeric item as integer.
}
\description{
Converts every numeric item in the list to integer.
}
