% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{pauseType}
\alias{pauseType}
\title{pauseType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
pauseType$new(temperature)
}
\arguments{
\item{temperature}{\link[base]{numeric}. The temperature in degrees Celsius
  to maintain during the pause.}
}
\description{
This step allows to pause at a certain temperature. It is typically the last
step in an amplification protocol.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

