% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{PCA}
\alias{PCA}
\title{Principal Component Analysis (PCA).}
\usage{
PCA(num_dims = NULL, thres = NULL)
}
\arguments{
\item{num_dims}{Number of components for dimensionality reduction. If NULL, all the principal components will be taken. Ignored if thres is provided. Integer.}

\item{Fraction}{of variability to keep, from 0 to 1. Data dimension will be reduced until the lowest dimension that keeps 'thres' explained variance. Float.}
}
\value{
The PCA transformer, structured as a named list.
}
\description{
A distance metric learning algorithm for unsupervised dimensionality reduction, obtaining orthogonal directions that maximize the variance.
}
