% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.RDML.R
\docType{methods}
\name{Plot}
\alias{Plot}
\alias{RDML.Plot}
\title{Plots \code{RDML} object}
\arguments{
\item{print.legend}{\code{logical} defines print or not plot legend}

\item{separate.by}{\code{list}}

\item{col}{sth}

\item{empty.col}{sth}

\item{...}{additional graphical parameters.}
}
\description{
Plots \code{RDML} objects.
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>, Stefan Roediger
  <stefan.roediger@hs-lausitz.de>, Michal Burdukiewicz
  <michalburdukiewicz@gmail.com>
}

