% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.RDML}
\alias{plot,RDML_object-method}
\alias{plot.RDML}
\alias{plot.RDML_object}
\title{Plot RDML_objects}
\usage{
plot.RDML(object, print.legend = TRUE, separate.by = list(left = c("name",
  "type", "targets"), right = c("name", "type", "targets")), col = list(left =
  NA, right = NA), empty.col = "white", ...)
}
\arguments{
\item{object}{an object of class \code{RDML}.}

\item{print.legend}{\code{logical}}

\item{separate.by}{\code{list}}

\item{col}{sth}

\item{empty.col}{sth}

\item{...}{additional graphical parameters.}
}
\description{
Plots \code{RDML} objects.
}

