% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.RDML_object}
\alias{plot,RDML_object-method}
\alias{plot.RDML_object}
\title{Plot RDML_objects}
\usage{
plot.RDML_object(x, print.legend = TRUE, separate.by = list(left = c("name",
  "type", "targets"), right = c("name", "type", "targets")), col = list(left =
  NA, right = NA), empty.col = "white", ...)
}
\arguments{
\item{x}{an object of class \code{RDML_object}.}

\item{print.legend}{\code{logical}}

\item{separate.by}{\code{list}}

\item{col}{sth}

\item{empty.col}{sth}

\item{...}{additional graphical parameters.}
}
\description{
Plots \code{RDML_object}s.
}

