% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{plotters}
\alias{plotters}
\title{Plot functions for Distance Metric Learning.}
\format{An object of class \code{character} of length 3.}
\usage{
plotters
}
\description{
Plot functions for Distance Metric Learning.
}
\keyword{datasets}
