% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{primingMethodType}
\alias{primingMethodType}
\title{primingMethodType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
primingMethodType$new(value)
}
\arguments{
\item{value}{\link[assertthat]{is.string}. Value.}
}
\description{
The primers used to reverse transcribe the RNA to cDNA. Can take values:
\describe{
\item{oligo-dt}{}
\item{random}{}
\item{target-specific}{}
\item{oligo-dt and random}{}
\item{other}{}
}
}
\details{
\cr Inherits: \link{enumType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

