% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDML.gui.R
\name{RDML.gui}
\alias{RDML.gui}
\title{Graphical User Interface for .rdml data import}
\usage{
RDML.gui()
}
\description{
\code{RDML.gui} is a web-browser based graphical interface allowing user to
read \code{.rdml} file, extract data and save it in an other format.
Additional features include summary statistics and plotting functions.
}
\section{Warning }{
 Any ad-blocking software may cause of malfunctions.
}
\examples{
\donttest{
library(shiny)
RDML.gui()
}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>, Stefan Roediger
<stefan.roediger@hs-lausitz.de>, Michal Burdukiewicz
<michalburdukiewicz@gmail.com>
}
\seealso{
Read \code{.rdml} files: \code{\link{RDML}}. Select data from
\code{.rdml} files: \code{\link{selectFData}}. Summary statistics:
\code{\link{summary.RDML_object}}.
}
\keyword{GUI}
\keyword{browser}
\keyword{hplot}
\keyword{shiny}

