\name{RDML_gui}
\alias{RDML_gui}

\title{
Graphical User Interface for .rdml data import
}
\description{
\code{RDML_gui} bla bla
}
\usage{
RDML_gui()
}

\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>, 
Stefan Roediger <stefan.roediger@hs-lausitz.de>,
Michal Burdukiewicz <michalburdukiewicz@gmail.com>
}
\section{Warning }{
Any ad-blocking software may be cause of malfunctions. 
}

\examples{
\donttest{
library(shiny)
RDML_gui()
}
}

\seealso{
\code{\link[chipPCR]{MFIaggr}}
}

\keyword{ hplot }
\keyword{ shiny }
\keyword{ GUI }
\keyword{ browser }