% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{rdmlBaseType}
\alias{rdmlBaseType}
\title{Base R6 class for RDML package.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Most classes from RDML package inherit this class. It is designed for internal
usage and should not be directly accessed.
}
\section{Initialization}{
 \preformatted{rdmlBaseType$new()}
}

\section{Methods}{
 \describe{
\item{\code{.asXMLnodes(node.name)}}{Represents
  object as XML nodes. Should not be called directly. \code{node.name} --
  name of the root node for the generated XML
  tree.}
\item{\code{print(...)}}{prints object}}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-.asXMLnodes}{\code{rdmlBaseType$.asXMLnodes()}}
\item \href{#method-copy}{\code{rdmlBaseType$copy()}}
\item \href{#method-print}{\code{rdmlBaseType$print()}}
\item \href{#method-clone}{\code{rdmlBaseType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.asXMLnodes"></a>}}
\if{latex}{\out{\hypertarget{method-.asXMLnodes}{}}}
\subsection{Method \code{.asXMLnodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rdmlBaseType$.asXMLnodes(node.name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\if{latex}{\out{\hypertarget{method-copy}{}}}
\subsection{Method \code{copy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rdmlBaseType$copy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rdmlBaseType$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rdmlBaseType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
