% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmlEdit.R
\name{rdmlEdit}
\alias{rdmlEdit}
\title{RDML Editor Graphical User Interface}
\usage{
rdmlEdit()
}
\description{
Launches graphical user interface that can edit RDML metadata and show
qPCR or melting curves.
}
\keyword{hplot}
