% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{rinitialize}
\alias{rinitialize}
\title{R DML objects initializer.}
\usage{
rinitialize(dml_alg)
}
\arguments{
\item{dml_alg}{The DML_Algorithm object imported form pyDML.}
}
\value{
a list with the DML_Algorithm methods adapted to R.
}
\description{
Initializes a DML as a list from the pyDML DML_Algorithm object.
}
