% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_clf.R
\name{rinitialize_clf}
\alias{rinitialize_clf}
\title{R classifiers initializer.}
\usage{
rinitialize_clf(clf)
}
\arguments{
\item{clf}{The classifier object imported form pyDML.}
}
\value{
a list with the classifier methods adapted to R.
}
\description{
Initializes a classifier as a list from a pyDML classifier.
}
