% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{sampleTypeType}
\alias{sampleTypeType}
\title{sampleTypeType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
sampleTypeType$new(value)
}
\arguments{
\item{value}{\link[assertthat]{is.string}. Value.}
}
\description{
Can take values:
\describe{
\item{unkn}{unknown sample}
\item{ntc}{non template control}
\item{nac}{no amplification control}
\item{std}{standard sample}
\item{ntp}{no target present}
\item{nrt}{minusRT}
\item{pos}{positive control}
\item{opt}{optical calibrator sample}}
}
\details{
\cr Inherits: \link{enumType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

