% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDML.Save.RDML.R
\docType{methods}
\name{Save}
\alias{RDML.Save}
\alias{Save}
\title{Saves \code{RDML} object as RDML v1.2 format file}
\arguments{
\item{file.name}{name of the generated RDML file}
}
\value{
\code{XML} representing \code{RDML} object as RDML v1.2 format
}
\description{
Saves \code{RDML} object as RDML v1.2 format file (if \code{file.name} is
specified) or returns XML tree as text.
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>, Stefan Roediger
  <stefan.roediger@hs-lausitz.de>, Michal Burdukiewicz
  <michalburdukiewicz@gmail.com>
}

