\name{selectFData}
\alias{selectFData}

\title{
Select fluorescence data vectors.
}
\description{
Selects fluorescence data vectors from \code{RDML_object} with specified types, targets and sample names.
}
\usage{
selectFData(object, melt = FALSE, targets = NA, types = NA, snames = NA)
}
\arguments{
  \item{object}{
  RDML_object.
  }
  \item{melt}{
  if TRUE -- use melting data.
  }
  \item{targets}{
  run targets to choose from (See 'Details').
  }
  \item{types}{
  sample types to choose from (See 'Details').
  }
  \item{snames}{
  regular expression to select sample names (See 'Details')
  }
}
\details{
If \code{targets = NA}, then use all of the available targets from run.\cr
If \code{types = NA}, then use all of the available types from specified \code{targets}.\cr
\code{snames} are really not samples names but columns names generated by \code{RDML name.pattern}.
\code{targets} and \code{types} options have higher priority than \code{snames} option. In other words if you specified \code{targets} or \code{types} you can't select samples with other type or target despite specified \code{snames}.\cr
}

\section{Warning}{
  Resulting \code{data.frame} can contain columns with same names! Use \code{name.pattern} option from \code{RDML} function to avoid this.
}

\value{
\code{data.frame} which contains selected fluorescence data \code{vector}s and Cycles/Temperatures in first column.
}


\references{
Regular Expressions as used in R -- http://stat.ethz.ch/R-manual/R-devel/library/base/html/regex.html
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>, 
Stefan Roediger <stefan.roediger@hs-lausitz.de>,
Michal Burdukiewicz <michalburdukiewicz@gmail.com>
}
\examples{
## EXAMPLE 1:
## internal dataset BioRad_qPCR_melt.rdml (in 'data' directory)
## generated by Bio-Rad CFX96. Contains qPCR and melting data.
## Import without splitting by targets/types and with
## custom name pattern.
library(chipPCR)
PATH <- path.package("RDML")
filename <- paste(PATH, "/extdata/", "BioRad_qPCR_melt.rdml", sep ="")
cfx96 <- RDML(filename, 
              name.pattern = "\%TUBE\%_\%NAME\%_\%TYPE\%_\%TARGET\%")
## Select samples with names that contain word 'unkn'.
fdata <- selectFData(cfx96, snames="unkn")
## Show names of selected samples
names(fdata)
## Use plotCurves function from the chiPCR package to 
## get an overview of the amplification curves
plotCurves(fdata[, 1], fdata[, -1])

## EXAMPLE 2:
## internal dataset lc96_bACTXY.rdml (in 'data' directory)
## generated by Roche LightCycler 96. Contains qPCR data
## with four targets and two types.
## Import with default settings.                     
PATH <- path.package("RDML")
filename <- paste(PATH, "/extdata/", "lc96_bACTXY.rdml", sep ="")
lc96 <- RDML(filename)
## Select samples with targets 'FAM@bACT' or 'Hex@X',
## type 'std' and names that contain word 'Sample 39' or 'Sample 4'
fdata <- selectFData(lc96, targets=c("FAM@bACT", "Hex@X"), types = "std", snames="(Sample 39)|(Sample 4)")
## Show names of selected samples
names(fdata)
}
\keyword{ manip }