% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{sequencesType}
\alias{sequencesType}
\title{sequencesType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
sequencesType
}
\description{
\cr Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \code{sequencesType$new(forwardPrimer = NULL, reversePrimer = NULL, probe1 = NULL, probe2 = NULL, amplicon = NULL)}
}

\section{Fields}{
 \describe{
\item{\code{forwardPrimer}}{\link{oligoType}.}
\item{\code{reversePrimer}}{\link{oligoType}.}
\item{\code{probe1}}{\link{oligoType}.}
\item{\code{probe2}}{\link{oligoType}.}
\item{\code{amplicon}}{\link{oligoType}.}
  }
}
\keyword{datasets}

