% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_wrappers.R
\name{SetFData}
\alias{SetFData}
\title{\code{RDML$SetFData()} wrapper}
\usage{
SetFData(obj, ...)
}
\arguments{
\item{obj}{\code{RDML} object.}

\item{...}{\code{SetFData} params.}
}
\description{
Read more at \link{RDML.SetFData}
}

