% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.RDML.R
\name{summary.RDML}
\alias{summary.RDML}
\title{S3 wrapper for \code{RDML$Summarize()} method}
\usage{
summary.RDML(object, print = TRUE, ...)
}
\description{
See \link{RDML.Summarize}.
}
\keyword{manip}
\keyword{summary}

