% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{supervised_dmls}
\alias{supervised_dmls}
\title{Supervised Distance Metric Learning algorithms.}
\format{An object of class \code{character} of length 16.}
\usage{
supervised_dmls
}
\description{
Supervised Distance Metric Learning algorithms.
}
\keyword{datasets}
