% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{targetType}
\alias{targetType}
\title{targetType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
targetType
}
\description{
A target is a PCR reaction with defined set of primers. PCR reactions
for the same gene with distinct primer sequences are considered different
targets. Inherits:
\link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{targetType$new(id, description = NULL,
  documentation = NULL, xRef = NULL, type, amplificationEfficiencyMethod =
  NULL, amplificationEfficiency = NULL, amplificationEfficiencySE = NULL,
  detectionLimit = NULL, dyeId, sequences = NULL, commercialAssay = NULL)}
}

\section{Fields}{
 \describe{ \item{\code{id}}{\link{idType}.}
  \item{\code{description}}{\link[checkmate]{checkString}.}
  \item{\code{documentation}}{\code{list} of \link{idReferencesType}.}
  \item{\code{xRef}}{\code{list} of \link{xRefType}.}
  \item{\code{type}}{\link{targetTypeType}.}
  \item{\code{amplificationEfficiencyMethod}}{\link[checkmate]{checkString}.}
  \item{\code{amplificationEfficiency}}{\link[checkmate]{checkNumber}.}
  \item{\code{amplificationEfficiencySE}}{\link[checkmate]{checkNumber}.}
  \item{\code{detectionLimit}}{\link[checkmate]{checkNumber}.}
  \item{\code{dyeId}}{\link{idReferencesType}.}
  \item{\code{sequences}}{\link{sequencesType}.}
  \item{\code{commercialAssay}}{\link{commercialAssayType}.} }
}
\keyword{datasets}

