% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{templateQuantityType}
\alias{templateQuantityType}
\title{templateQuantityType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
templateQuantityType$new(conc, nucleotide)
}
\arguments{
\item{conc}{\link[assertthat]{is.float}. Concentration of the template in nanogram per microliter in the final reaction mix.}

\item{nucleotide}{\link{nucleotideType}.}
}
\description{
\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

