% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{Transformer}
\alias{Transformer}
\title{Basic learning from a transformer.}
\usage{
Transformer(transformer)
}
\arguments{
\item{metric}{d' x d) matrix, representing a linear transformacion from d-dimensional euclidean space
to d'-dimensional euclidean space.}
}
\value{
The Transformer transformer, structured as a named list.
}
\description{
A DML algorithm that defines a distance given a linear transformation.
}
