% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{tuners}
\alias{tuners}
\title{Distance Metric Learning tune algorithms.}
\format{An object of class \code{character} of length 2.}
\usage{
tuners
}
\description{
Distance Metric Learning tune algorithms.
}
\keyword{datasets}
