% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.Write.R
\docType{methods}
\name{Write}
\alias{RDML.Write}
\alias{Write}
\title{Write \code{RDML} objects}
\description{
Writes \code{RDML} object.
}
\examples{
\dontrun{
PATH <- path.package("RDML")
filename <- paste(PATH, "/extdata/", "lc96_bACTXY.rdml", sep ="")
lc96 <- RDML$new(filename)
filename <- paste(PATH, "/extdata/", "stepone_std.rdml", sep ="")
stepone <- RDML$new(filename)
lc96$Merge(stepone)
lc96$AsTable()
}
}

