% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{xRefType}
\alias{xRefType}
\title{xRefType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
xRefType
}
\description{
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{xRefType$new(name = NULL, id = NULL)}
}

\section{Fields}{
 \describe{
  \item{\code{name}}{\link[assertthat]{is.string}. Reference to an external
  database, } for example "GenBank".
  \item{\code{id}}{\link[assertthat]{is.string}. The ID of the entry within
  the external database, for example "AJ832138".}
  }
}
\keyword{datasets}

