% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{annotationType}
\alias{annotationType}
\title{annotationType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
annotationType$new(property, value)
}
\arguments{
\item{property}{\link[assertthat]{is.string}. Property}

\item{value}{\link[assertthat]{is.string}. Value}
}
\description{
These elements should be used to annotate samples by setting a property and a
value. A property could be sex, the value M or F.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

