% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{annotationType}
\alias{annotationType}
\title{annotationType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
annotationType
}
\description{
Annotate samples by setting a property and its value. For example,
sex could be a property with the possible values M or F. Inherits:
\link{rdmlBaseType}.
}
\section{Fields}{
 \describe{ \item{property}{\link[assertthat]{is.string}.
  Property name} \item{value}{\link[assertthat]{is.string}. Value} }
}
\examples{
#set sex property
annotationType$new(property = "sex", value = "M")
}
\keyword{datasets}

