% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{annotationType}
\alias{annotationType}
\title{annotationType R6 class.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Annotate samples by setting a property and its value. For example,
sex could be a property with the possible values M or F. Inherits:
\link{rdmlBaseType}.
}
\section{Fields}{
 \describe{ \item{property}{\link[checkmate]{checkString}.
  Property name} \item{value}{\link[checkmate]{checkString}. Value} }
}

\examples{
#set sex property
annotationType$new(property = "sex", value = "M")
}
\section{Super class}{
\code{\link[RDML:rdmlBaseType]{RDML::rdmlBaseType}} -> \code{annotationType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-annotationType-new}{\code{annotationType$new()}}
\item \href{#method-annotationType-clone}{\code{annotationType$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id=".asXMLnodes"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-.asXMLnodes'><code>RDML::rdmlBaseType$.asXMLnodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="copy"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-copy'><code>RDML::rdmlBaseType$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="print"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-print'><code>RDML::rdmlBaseType$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotationType-new"></a>}}
\if{latex}{\out{\hypertarget{method-annotationType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{annotationType$new(property, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotationType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-annotationType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{annotationType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
