% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.AsDendrogram.R
\docType{methods}
\name{AsDendrogram}
\alias{AsDendrogram}
\alias{RDML.AsDendrogram}
\title{Represents structure of \code{RDML} file as dendrogram}
\arguments{
\item{plot.dendrogram}{plots dendrogram if \code{TRUE}}
}
\value{
\code{dendrogram} object
}
\description{
Plots and/or returns structure of \code{RDML} file as dendrogram (tree view)
}
\examples{
\dontrun{
PATH <- path.package("RDML")
filename <- paste(PATH, "/extdata/", "BioRad_qPCR_melt.rdml", sep ="")
cfx96 <- RDML$new(filename)
cfx96$AsDendrogram()
}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>, Stefan Roediger
  <stefan.roediger@b-tu.de>, Michal Burdukiewicz
  <michalburdukiewicz@gmail.com>
}
\keyword{manip}

