% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{baseTemperatureType}
\alias{baseTemperatureType}
\title{baseTemperatureType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
baseTemperatureType
}
\description{
Parent class for inner usage. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{baseTemperatureType$new(duration,
  temperatureChange = NULL, durationChange = NULL, measure = NULL, ramp =
  NULL)}
}

\section{Fields}{
 \describe{
  \item{\code{duration}}{\link[checkmate]{checkCount}. Duration of this
  step in seconds.}
  \item{\code{temperatureChange}}{\link[checkmate]{checkNumber}. Change
  of the temperature between two consecutive cycles: actual temperature
  = temperature + (temperatureChange * cycle counter)}
  \item{\code{durationChange}}{\link[checkmate]{checkCount}. Change of the
  duration between two consecutive cycles: actual duration = duration +
  (durationChange * cycle counter)}
  \item{\code{measure}}{\link{measureType}. Indicates to make a measurement
  and store it as meltcurve or real-time data.}
  \item{\code{ramp}}{\link[checkmate]{checkNumber}. Allowed temperature
  change between two consecutive cycles in degrees Celsius per second. If unstated,
  the maximal change rate is assumed.}
  }
}
\keyword{datasets}

