% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{baseTemperatureType}
\alias{baseTemperatureType}
\title{baseTemperatureType R6 class.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Parent class for inner usage. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{baseTemperatureType$new(duration,
  temperatureChange = NULL, durationChange = NULL, measure = NULL, ramp =
  NULL)}
}

\section{Fields}{
 \describe{
  \item{\code{duration}}{\link[checkmate]{checkCount}. Duration of this
  step in seconds.}
  \item{\code{temperatureChange}}{\link[checkmate]{checkNumber}. Change
  of the temperature between two consecutive cycles: actual temperature
  = temperature + (temperatureChange * cycle counter)}
  \item{\code{durationChange}}{\link[checkmate]{checkCount}. Change of the
  duration between two consecutive cycles: actual duration = duration +
  (durationChange * cycle counter)}
  \item{\code{measure}}{\link{measureType}. Indicates to make a measurement
  and store it as meltcurve or real-time data.}
  \item{\code{ramp}}{\link[checkmate]{checkNumber}. Allowed temperature
  change between two consecutive cycles in degrees Celsius per second. If unstated,
  the maximal change rate is assumed.}
  }
}

\section{Super class}{
\code{\link[RDML:rdmlBaseType]{RDML::rdmlBaseType}} -> \code{baseTemperatureType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{baseTemperatureType$new()}}
\item \href{#method-clone}{\code{baseTemperatureType$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id=".asXMLnodes">}\href{../../RDML/html/rdmlBaseType.html#method-.asXMLnodes}{\code{RDML::rdmlBaseType$.asXMLnodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="copy">}\href{../../RDML/html/rdmlBaseType.html#method-copy}{\code{RDML::rdmlBaseType$copy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="print">}\href{../../RDML/html/rdmlBaseType.html#method-print}{\code{RDML::rdmlBaseType$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baseTemperatureType$new(
  duration,
  temperatureChange = NULL,
  durationChange = NULL,
  measure = NULL,
  ramp = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baseTemperatureType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
