% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{cdnaSynthesisMethodType}
\alias{cdnaSynthesisMethodType}
\title{cdnaSynthesisMethodType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
cdnaSynthesisMethodType$new(enzyme = NULL, primingMethod = NULL, dnaseTreatment = NULL, thermalCyclingConditions = NULL)
}
\arguments{
\item{enzyme}{\link[assertthat]{is.float}. Enzyme used for reverse transcription.}

\item{primingMethod}{\link{primingMethodType}.}

\item{dnaseTreatment}{\link[assertthat]{is.flag} True if RNA was DNAse treated prior cDNA synthesis.}

\item{thermalCyclingConditions}{\link{idReferencesType}.}
}
\description{
Description of the cDNA synthesis method.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

