% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{cdnaSynthesisMethodType}
\alias{cdnaSynthesisMethodType}
\title{cdnaSynthesisMethodType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
cdnaSynthesisMethodType
}
\description{
Description of the cDNA synthesis method. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{cdnaSynthesisMethodType$new(enzyme = NULL,
  primingMethod = NULL, dnaseTreatment = NULL, thermalCyclingConditions =
  NULL)}

  @section Fields: \describe{
  \item{\code{enzyme}}{\link[checkmate]{checkString}. Name of the enzyme used for
  reverse transcription.}
  \item{\code{primingMethod}}{\link{primingMethodType}.}
  \item{\code{dnaseTreatment}}{\link[checkmate]{checkFlag} if \code{TRUE}RNA was
  DNAse treated prior cDNA synthesis.}
  \item{\code{thermalCyclingConditions}}{\link{idReferencesType}.}
  }
}

\keyword{datasets}
