% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{commercialAssayType}
\alias{commercialAssayType}
\title{commercialAssayType R6 class.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
For some commercial assays, the primer sequences may be unknown. This element
allows to describe commercial assays. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{commercialAssayType$new(company, orderNumber)}

  @section Fields: \describe{
\item{\code{company}}{\link[checkmate]{checkString}.}
\item{\code{orderNumber}}{\link[checkmate]{checkString}.}
  }
}

\section{Super class}{
\code{\link[RDML:rdmlBaseType]{RDML::rdmlBaseType}} -> \code{commercialAssayType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-commercialAssayType-new}{\code{commercialAssayType$new()}}
\item \href{#method-commercialAssayType-clone}{\code{commercialAssayType$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id=".asXMLnodes"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-.asXMLnodes'><code>RDML::rdmlBaseType$.asXMLnodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="copy"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-copy'><code>RDML::rdmlBaseType$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="print"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-print'><code>RDML::rdmlBaseType$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commercialAssayType-new"></a>}}
\if{latex}{\out{\hypertarget{method-commercialAssayType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commercialAssayType$new(company, orderNumber)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commercialAssayType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-commercialAssayType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commercialAssayType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
