% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{commercialAssayType}
\alias{commercialAssayType}
\title{commercialAssayType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
commercialAssayType
}
\description{
For some commercial assays, the primer sequences may be unknown. This element
allows to describe commercial assays. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{commercialAssayType$new(company, orderNumber)}

  @section Fields: \describe{
\item{\code{company}}{\link[checkmate]{checkString}.}
\item{\code{orderNumber}}{\link[checkmate]{checkString}.}
  }
}

\keyword{datasets}
