% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{dataCollectionSoftwareType}
\alias{dataCollectionSoftwareType}
\title{dataCollectionSoftwareType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
dataCollectionSoftwareType
}
\description{
Software name and version used to collect and analyze the data. Inherits:
\link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{dataCollectionSoftwareType$new(name, version)}
}

\section{Fields}{
 \describe{
\item{\code{name}}{\link[assertthat]{is.string}.}
\item{\code{version}}{\link[assertthat]{is.string}.}
  }
}
\examples{
dataCollectionSoftwareType$new(name = "ExampleSoft",
                               version = "1.0")
}
\keyword{datasets}

