% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{dataCollectionSoftwareType}
\alias{dataCollectionSoftwareType}
\title{dataCollectionSoftwareType R6 class.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Software name and version used to collect and analyze the data. Inherits:
\link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{dataCollectionSoftwareType$new(name, version)}

  @section Fields: \describe{
\item{\code{name}}{\link[checkmate]{checkString}.}
\item{\code{version}}{\link[checkmate]{checkString}.}
  }
}

\examples{
dataCollectionSoftwareType$new(name = "ExampleSoft",
                               version = "1.0")
}
\section{Super class}{
\code{\link[RDML:rdmlBaseType]{RDML::rdmlBaseType}} -> \code{dataCollectionSoftwareType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataCollectionSoftwareType-new}{\code{dataCollectionSoftwareType$new()}}
\item \href{#method-dataCollectionSoftwareType-clone}{\code{dataCollectionSoftwareType$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id=".asXMLnodes"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-.asXMLnodes'><code>RDML::rdmlBaseType$.asXMLnodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="copy"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-copy'><code>RDML::rdmlBaseType$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="print"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-print'><code>RDML::rdmlBaseType$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataCollectionSoftwareType-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataCollectionSoftwareType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataCollectionSoftwareType$new(name, version)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataCollectionSoftwareType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataCollectionSoftwareType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataCollectionSoftwareType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
