% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{dataType}
\alias{dataType}
\title{dataType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
dataType$new(tar, cq = NULL, excl = NULL, adp = NULL, mdp = NULL,
  endPt = NULL, bgFluor = NULL, bgFluorSlp = NULL, quantFluor = NULL)
}
\arguments{
\item{tar}{\link{idReferencesType}. TargetID - A reference to a target.}

\item{cq}{\link[base]{double}. Quantification cycle - The calculated
fractional PCR cycle used for downstream quantification. Negative values
are used to express following conditions: Not Available: -1.0}

\item{excl}{\link[assertthat]{is.string}. Excluded - If present, this entry
should not be evaluated. Do not set this element to false if this entry is
valid, leave the entire element out instead. It may contain a string with
reason for exclusion. Several reasons for exclusion should be seperated by
semicolons ";".}

\item{adp}{\link{adpsType}.}

\item{mdp}{\link{mdpsType}.}

\item{endPt}{\link[base]{double}. End point - Result of an endpoint
measurement.}

\item{bgFluor}{\link[base]{double}. Background fluorescence - The y-intercept
of the baseline trend based on the estimated background fluorescence.}

\item{bgFluorSlp}{\link[base]{double}. Background fluorescence slope - The
slope of the baseline trend based on the estimated background fluorescence.
The element should be absent to indicate a slope of 0.0; If this element is
present without the bgFluor element it should be ignored.}

\item{quantFluor}{\link[base]{double}. Quantification flourescence - The
  fluorescence value corresponding to the treshold line.}
}
\description{
\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}

\section{Methods}{
 \describe{\item{\code{AsDataFrame(dp.type =
  "adp")}}{Represents amplification (\code{dp.type = "adp"}) or melting
  (\code{dp.type = "mdp"}) data points as \code{data.frame}}}
}
\keyword{datasets}

