% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{dyeType}
\alias{dyeType}
\title{dyeType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
dyeType$new(id, description = NULL)
}
\arguments{
\item{id}{\link{idType}. Identificator.}

\item{text}{\link[assertthat]{is.string}. Description.}
}
\description{
Information on a dye.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

