% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{dyeType}
\alias{dyeType}
\title{dyeType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
dyeType
}
\description{
Detailed information about the dye. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{dyeType$new(id, description = NULL)}

  @section Fields: \describe{
\item{\code{id}}{\link{idType}. Identificator.}
\item{\code{description}}{ \link[checkmate]{checkString}. Description.
  }}
}
\keyword{datasets}

