% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{dyeType}
\alias{dyeType}
\title{dyeType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
dyeType
}
\description{
Information on a dye.Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{dyeType$new(id, description = NULL)}
}

\section{Fields}{
 \describe{
\item{\code{id}}{\link{idType}. Identificator.}
\item{\code{description}}{ \link[assertthat]{is.string}. Description.
  }}
}
\keyword{datasets}

