% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{enumType}
\alias{enumType}
\title{enumType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
enumType$new(value)
}
\arguments{
\item{value}{\link[assertthat]{is.string}. Value.}
}
\description{
Generic class for creating objects thet can take limited list of values.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

