% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{enumType}
\alias{enumType}
\title{enumType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
enumType
}
\description{
Generic class for creating objects thet can take limited list of values. \cr
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{enumType$new(value)}
  @section Fields: \describe{
\item{\code{value}}{\link[checkmate]{checkString}. Value.}
  }
}

\keyword{datasets}
