% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{experimenterType}
\alias{experimenterType}
\title{experimenterType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
experimenterType
}
\description{
Contact details of the experimenter. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{experimenterType$new(id, firstName, lastName,
  email = NULL, labName = NULL, labAddress = NULL)}

  @section Fields: \describe{
\item{\code{id}}{\link{idType}. Identificator.}
\item{\code{firstName}}{\link[checkmate]{checkString}. First name.}
\item{\code{lastName}}{\link[checkmate]{checkString}. Last name.}
\item{\code{email}}{\link[checkmate]{checkString}. Email.}
\item{\code{labName}}{\link[checkmate]{checkString}. Lab name.}
\item{\code{labAddress}}{\link[checkmate]{checkString}. Lab address.}
  }
}
\keyword{datasets}

