% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{experimenterType}
\alias{experimenterType}
\title{experimenterType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
experimenterType
}
\description{
Contact details of the experimenter.Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{experimenterType$new(id, firstName, lastName,
  email = NULL, labName = NULL, labAddress = NULL)}
}

\section{Fields}{
 \describe{
\item{\code{id}}{\link{idType}. Identificator.}
\item{\code{firstName}}{\link[assertthat]{is.string}. First name.}
\item{\code{lastName}}{\link[assertthat]{is.string}. Last name.}
\item{\code{email}}{\link[assertthat]{is.string}. Email.}
\item{\code{labName}}{\link[assertthat]{is.string}. Lab name.}
\item{\code{labAddress}}{\link[assertthat]{is.string}. Lab address.}
  }
}
\keyword{datasets}

