% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{gradientType}
\alias{gradientType}
\title{gradientType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
gradientType$new(highTemperature, lowTemperature, ...)
}
\arguments{
\item{highTemperature}{\link[base]{double}. The high temperature of the
gradient in degrees Celsius.}

\item{lowTemperature}{\link[base]{double}. The low temperature of the
gradient in degrees Celsius.}

\item{...}{Params of parent class \link{baseTemperatureType}.}
}
\description{
This step forms a temperature gradient across the PCR block. Inherits:
\link{baseTemperatureType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

