% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{gradientType}
\alias{gradientType}
\title{gradientType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
gradientType
}
\description{
This step forms a temperature gradient across the PCR block. Inherits:
\link{baseTemperatureType}.
}
\section{Initialization}{
 \preformatted{gradientType$new(highTemperature,
  lowTemperature, ...)}
}

\section{Fields}{
 \describe{
  \item{\code{highTemperature}}{\link[base]{double}. The high temperature of
  thegradient in degrees Celsius.}
  \item{\code{lowTemperature}}{\link[base]{double}. The low temperature of
  the gradient in degrees Celsius.}
\item{\code{...}}{ Params of parent class \link{baseTemperatureType}. }}
}
\keyword{datasets}

