% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{idReferencesType}
\alias{idReferencesType}
\title{idReferencesType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
idReferencesType$new(id)
}
\arguments{
\item{id}{\link[assertthat]{is.string}. Identificator.}
}
\description{
Contains id of another RDML object.\cr Inherits: \link{idType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

