% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{idReferencesType}
\alias{idReferencesType}
\title{idReferencesType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
idReferencesType
}
\description{
Contains id of another RDML object.Inherits: \link{idType}.
}
\section{Initialization}{
 \code{idReferencesType$new(id)}
}

\section{Fields}{
 \describe{
\item{\code{id}}{\link[assertthat]{is.string}. Identificator.}
}
}
\keyword{datasets}

