% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{idType}
\alias{idType}
\title{idType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
idType$new(id)
}
\arguments{
\item{id}{\link[assertthat]{is.string}. Identificator.}
}
\description{
Contains identificator for varius RDML types.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

