% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{idType}
\alias{idType}
\title{idType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
idType
}
\description{
Contains identificator for varius RDML types.Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \code{idType$new(id)}
}

\section{Fields}{
 \describe{
\item{\code{id}}{\link[assertthat]{is.string}. Identificator.}
}
}
\keyword{datasets}

