% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{labelFormatType}
\alias{labelFormatType}
\title{labelFormatType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
labelFormatType$new(value)
}
\arguments{
\item{value}{\link[assertthat]{is.string}.}
}
\description{
Label used for \link{pcrFormat}.
Can take values:
\describe{
\item{ABC}{}
\item{123}{}
\item{A1a1}{}
}
\cr Inherits: \link{enumType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

