% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{lidOpenType}
\alias{lidOpenType}
\title{lidOpenType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
lidOpenType
}
\description{
This step waits for the user to open the lid and continues afterwards. It
allows to stop the program and to wait for the user to add for example
enzymes and continue the program afterwards. The temperature of the previous
step is maintained. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{lidOpenType$new()}
}
\keyword{datasets}

