% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{mdpsType}
\alias{mdpsType}
\title{mdpsType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
mdpsType
}
\description{
Contains \code{matrix} of melting data points (single data points measured
during amplification).
}
\details{
Columns: \describe{
\item{tmp}{(temperature in degrees Celsius at the time of measurement.
Every point must have unique value.}
\item{fluor}{fluorescence intensity measured without any correction
(including baselining).}}

Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{mdpsType$new(fpoints)}
}

\section{Fields}{
 \describe{
\item{\code{fpoints}}{\link[base]{matrix}. Matrix with amplification data points.}
  }
}
\keyword{datasets}

