% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{mdpsType}
\alias{mdpsType}
\title{mdpsType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
mdpsType
}
\description{
Contains of melting data points \code{matrix} -- single data points measured
during amplification. \code{Matrix} columns: \describe{ \item{tmp}{(every
point must be unique) Temperature - The temperature in degrees Celsius at the
time of measurement.} \item{fluor}{Fluorescence - The fluorescence intensity
measured without any correction. The fluorescence intensity must not be
baseline corrected.}} Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{mdpsType$new(fpoints)}
}

\section{Fields}{
 \describe{
\item{\code{fpoints}}{\link[base]{matrix}. Matrix with amplification data points.}
  }
}
\keyword{datasets}

