% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDML.Merge.R
\docType{methods}
\name{Merge}
\alias{Merge}
\alias{RDML.Merge}
\title{Merges \code{RDML} objects}
\arguments{
\item{to.merge}{\code{RDML} object that should be merged into current object.}
}
\description{
Merges one \code{RDML} object into another.
}
\examples{
\dontrun{
PATH <- path.package("RDML")
filename <- paste(PATH, "/extdata/", "lc96_bACTXY.rdml", sep ="")
lc96 <- RDML$new(filename)
filename <- paste(PATH, "/extdata/", "stepone_std.rdml", sep ="")
stepone <- RDML$new(filename)
lc96$Merge(stepone)
lc96$AsTable()
}
}

