% Generated by roxygen2 (4.0.2): do not edit by hand
\name{names.RDML}
\alias{names,RDML-method}
\alias{names.RDML}
\title{The Names of the RDML object}
\usage{
\method{names}{RDML}(object)
}
\arguments{
\item{object}{an object of class \code{RDML}.}
}
\description{
Just wraps \code{ls()} function to show RDML public fields
and methods. Useful for autocomplete at R editors.
}

