% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{nucleotideType}
\alias{nucleotideType}
\title{nucleotideType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
nucleotideType
}
\description{
Type of nucleic acid used as a template in the experiment. May have following values:
\describe{
\item{DNA}{}
\item{genomic-DNA}{}
\item{cDNA}{}
\item{RNA}{}
}
}
\details{
Inherits: \link{enumType}.
}
\section{Initialization}{
 \preformatted{nucleotideType$new(value)}
}

\section{Fields}{
 \describe{
\item{\code{value}}{\link[assertthat]{is.string}. Value.}
  }
}
\keyword{datasets}

