% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{oligoType}
\alias{oligoType}
\title{oligoType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
oligoType$new(threePrimeTag = NULL, fivePrimeTag = NULL, sequence)
}
\arguments{
\item{threePrimeTag}{\link[assertthat]{is.string}. Description of three prime modification (if present).}

\item{fivePrimeTag}{\link[assertthat]{is.string}. Description of five prime modification (if present).}

\item{sequence}{\link[assertthat]{is.string}.}
}
\description{
\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

