% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{oligoType}
\alias{oligoType}
\title{oligoType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
oligoType
}
\description{
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \code{oligoType$new(threePrimeTag = NULL,
  fivePrimeTag = NULL, sequence)}
}

\section{Fields}{
 \describe{
  \item{\code{threePrimeTag}}{\link[assertthat]{is.string}. Description of
  three prime modification (if present). }
  \item{\code{fivePrimeTag}}{\link[assertthat]{is.string}. Description of
  five prime modification (if present).}
  \item{\code{sequence}}{\link[assertthat]{is.string}.}
  }
}
\keyword{datasets}

