% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{oligoType}
\alias{oligoType}
\title{oligoType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
oligoType
}
\description{
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{oligoType$new(threePrimeTag = NULL,
  fivePrimeTag = NULL, sequence)}

  @section Fields: \describe{
  \item{\code{threePrimeTag}}{\link[checkmate]{checkString}. Description of
  three prime modification (if present). }
  \item{\code{fivePrimeTag}}{\link[checkmate]{checkString}. Description of
  five prime modification (if present).}
  \item{\code{sequence}}{\link[checkmate]{checkString}.}
  }
}

\keyword{datasets}
