% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{pauseType}
\alias{pauseType}
\title{pauseType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
pauseType
}
\description{
This step allows to pause at a certain temperature. It is typically the last
step in an amplification protocol.Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \code{pauseType$new(temperature)}
}

\section{Fields}{
 \describe{ \item{\code{temperature}}{\link[base]{numeric}.
  The temperature in degrees Celsius to maintain during the pause.}
  }
}
\keyword{datasets}

