% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.RDML}
\alias{plot.RDML}
\title{S3 wrapper for \code{RDML$Plot()} method}
\usage{
\method{plot}{RDML}(object, print.legend = TRUE, separate.by = list(left =
  c("name", "type", "targets"), right = c("name", "type", "targets")),
  col = list(left = NA, right = NA), empty.col = "white", ...)
}
\description{
See \link{RDML.Plot}.
}

