% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.RDML.R
\name{plot.RDML}
\alias{plot.RDML}
\title{S3 wrapper for \code{RDML$Plot()} method}
\usage{
plot.RDML(object, print.legend = TRUE, separate.by = list(left = c("name",
  "type", "targets"), right = c("name", "type", "targets")), col = list(left =
  NA, right = NA), empty.col = "white", ...)
}
\description{
See \link{RDML.Plot}.
}

