% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{primingMethodType}
\alias{primingMethodType}
\title{primingMethodType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
primingMethodType
}
\description{
The primers used to reverse transcribe the RNA to cDNA. Can take values:
\describe{
\item{oligo-dt}{}
\item{random}{}
\item{target-specific}{}
\item{oligo-dt and random}{}
\item{other}{}
}
}
\details{
\cr Inherits: \link{enumType}.
}
\section{Initialization}{
 \code{primingMethodType$new(value)}
}

\section{Fields}{
 \describe{
\item{\code{value}}{\link[assertthat]{is.string}. Value.}
  }
}
\keyword{datasets}

