% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{primingMethodType}
\alias{primingMethodType}
\title{primingMethodType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
primingMethodType
}
\description{
The primers used in the reverse transcription. Can take values:
\describe{
\item{oligo-dt}{}
\item{random}{}
\item{target-specific}{}
\item{oligo-dt and random}{}
\item{other}{}
}
}
\details{
Inherits: \link{enumType}.
}
\section{Initialization}{
 \preformatted{primingMethodType$new(value)}

  @section Fields: \describe{
\item{\code{value}}{\link[checkmate]{checkString}. Value.}
  }
}

\keyword{datasets}
