% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{quantityType}
\alias{quantityType}
\title{quantityType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
quantityType$new(value, unit)
}
\arguments{
\item{value}{\link[assertthat]{is.float}. Value.}

\item{unit}{\link{quantityUnitType}. Unit.}
}
\description{
A quantity is always defined by its value and its unit.\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

