% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{quantityType}
\alias{quantityType}
\title{quantityType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
quantityType
}
\description{
A quantity is always defined by its value and its unit. Inherits:
\link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{quantityType$new(value, unit)}

  @section Fields: \describe{
\item{\code{value}}{\link[checkmate]{checkNumber}. Value.}
\item{\code{unit}}{\link{quantityUnitType}. Unit.}
  }
}

\keyword{datasets}
