% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{quantityUnitType}
\alias{quantityUnitType}
\title{quantityUnitType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
quantityUnitType
}
\description{
The unit the quantity. Can take values:
\describe{
\item{cop}{copies per microliter  }
\item{fold}{fold change  }
\item{dil}{dilution (10 would mean 1:10 dilution)  }
\item{nMol}{nanomol per microliter }
\item{ng}{nanogram per microliter }
\item{other}{other unit (must be linear, no exponents or logarithms allowed) }
}
}
\details{
Inherits: \link{enumType}.
}
\section{Initialization}{
 \preformatted{quantityUnitType$new(value)}

  @section Fields: \describe{
\item{\code{value}}{\link[checkmate]{checkString}. Value.}
  }
}

\keyword{datasets}
