% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{rdmlBaseType}
\alias{rdmlBaseType}
\title{Base R6 class for RDML package.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
rdmlBaseType
}
\description{
Most classes from RDML package inherit this class. It is designed for internal
usage and should not be directly accessed.
}
\section{Initialization}{
 \preformatted{rdmlBaseType$new()}
}

\section{Methods}{
 \describe{
\item{\code{.asXMLnodes(node.name)}}{Represents
  object as XML nodes. Should not be called directly. \code{node.name} --
  name of the root node for the generated XML
  tree.}
\item{\code{print(...)}}{prints object}}
}

\keyword{datasets}
