% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{rdmlIdType}
\alias{rdmlIdType}
\title{rdmlIdType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
rdmlIdType
}
\description{
This element can be used to assign a publisher and id to the RDML file.\cr 
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{rdmlIdType$new(publisher, serialNumber,
  MD5Hash = NULL)}
}

\section{Fields}{
 \describe{  
  \item{\code{publisher}}{\link[checkmate]{checkString}. RDML file publisher.}
  \item{\code{serialNumber}}{\link[checkmate]{checkString}. Serial number.} 
  \item{\code{MD5Hash}}{\link[checkmate]{checkString}. An MD5Hash calculated
  over the complete file after removing all rdmlIDTypes and all whitespaces
  between elements.}
  }
}
\keyword{datasets}

