% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{reactIdType}
\alias{reactIdType}
\title{reactIdType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
reactIdType
}
\description{
Contains identificator for reactType.Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \code{reactIdType$new(id)}
}

\section{Fields}{
 \describe{
\item{\code{id}}{\link[assertthat]{is.count}. Identificator.}
}
}
\keyword{datasets}

