% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{sampleTypeType}
\alias{sampleTypeType}
\title{sampleTypeType R6 class.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Can take values:
\describe{
\item{unkn}{unknown sample}
\item{ntc}{non template control}
\item{nac}{no amplification control}
\item{std}{standard sample}
\item{ntp}{no target present}
\item{nrt}{minusRT}
\item{pos}{positive control}
\item{opt}{optical calibrator sample}}
}
\details{
Inherits: \link{enumType}.
}
\section{Initialization}{
 \preformatted{sampleTypeType$new(value)}

  @section Fields: \describe{
\item{\code{value}}{\link[checkmate]{checkString}. Value.}
  }
}

\section{Super classes}{
\code{\link[RDML:rdmlBaseType]{RDML::rdmlBaseType}} -> \code{\link[RDML:enumType]{RDML::enumType}} -> \code{sampleTypeType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sampleTypeType-clone}{\code{sampleTypeType$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="copy"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-copy'><code>RDML::rdmlBaseType$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="enumType" data-id=".asXMLnodes"><a href='../../RDML/html/enumType.html#method-enumType-.asXMLnodes'><code>RDML::enumType$.asXMLnodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="enumType" data-id="initialize"><a href='../../RDML/html/enumType.html#method-enumType-initialize'><code>RDML::enumType$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="enumType" data-id="print"><a href='../../RDML/html/enumType.html#method-enumType-print'><code>RDML::enumType$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sampleTypeType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sampleTypeType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sampleTypeType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
