% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{sampleTypeType}
\alias{sampleTypeType}
\title{sampleTypeType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
sampleTypeType
}
\description{
Can take values:
\describe{
\item{unkn}{unknown sample}
\item{ntc}{non template control}
\item{nac}{no amplification control}
\item{std}{standard sample}
\item{ntp}{no target present}
\item{nrt}{minusRT}
\item{pos}{positive control}
\item{opt}{optical calibrator sample}}
}
\details{
Inherits: \link{enumType}.
}
\section{Initialization}{
 \preformatted{sampleTypeType$new(value)}
 
  @section Fields: \describe{   
\item{\code{value}}{\link[checkmate]{checkString}. Value.}
  }
}
\keyword{datasets}

