% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Save}
\alias{RDML.Save}
\alias{Save}
\title{Saves \code{RDML} object as RDML v1.2 format file}
\arguments{
\item{file.name}{name of the generated RDML file}
}
\value{
\code{XML} representing \code{RDML} object as RDML v1.2 format
}
\description{
Saves \code{RDML} object as RDML v1.2 format file (if \code{file.name} is
specified) or returns XML tree as text.
}

