% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{sequencesType}
\alias{sequencesType}
\title{sequencesType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
sequencesType$new(forwardPrimer = NULL, reversePrimer = NULL, probe1 =
  NULL, probe2 = NULL, amplicon = NULL)
}
\arguments{
\item{forwardPrimer}{\link{oligoType}.}

\item{reversePrimer}{\link{oligoType}.}

\item{probe1}{\link{oligoType}.}

\item{probe2}{\link{oligoType}.}

\item{amplicon}{\link{oligoType}.}
}
\description{
\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

