% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{stepType}
\alias{stepType}
\title{stepType R6 class.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{stepType$new(nr, description = NULL,
  temperature = NULL, gradient = NULL, loop = NULL, pause = NULL, lidOpen =
  NULL)}
}

\section{Fields}{
 \describe{ \item{\code{nr}}{\link[checkmate]{checkCount}. The
  incremental number of the step. First step should have value 1. The increment
  between steps should be constant and equivalent to 1.}
  \item{\code{description}}{\link[checkmate]{checkString}.}
  \item{\code{temperature}}{\link{temperatureType}.}
  \item{\code{gradient}}{\link{gradientType}.}
  \item{\code{loop}}{\link{loopType}.} \item{\code{pause}}{\link{pauseType}.}
  \item{\code{lidOpen}}{\link{lidOpenType}.}
  }
}

\section{Super class}{
\code{\link[RDML:rdmlBaseType]{RDML::rdmlBaseType}} -> \code{stepType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-stepType-new}{\code{stepType$new()}}
\item \href{#method-stepType-clone}{\code{stepType$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id=".asXMLnodes"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-.asXMLnodes'><code>RDML::rdmlBaseType$.asXMLnodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="copy"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-copy'><code>RDML::rdmlBaseType$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="print"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-print'><code>RDML::rdmlBaseType$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stepType-new"></a>}}
\if{latex}{\out{\hypertarget{method-stepType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stepType$new(
  nr,
  description = NULL,
  temperature = NULL,
  gradient = NULL,
  loop = NULL,
  pause = NULL,
  lidOpen = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stepType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-stepType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stepType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
