% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{stepType}
\alias{stepType}
\title{stepType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
stepType
}
\description{
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{stepType$new(nr, description = NULL,
  temperature = NULL, gradient = NULL, loop = NULL, pause = NULL, lidOpen =
  NULL)}
}

\section{Fields}{
 \describe{ \item{\code{nr}}{\link[checkmate]{checkCount}. The
  incremental number of the step. First step should have value 1. The increment
  between steps should be constant and equivalent to 1.}
  \item{\code{description}}{\link[checkmate]{checkString}.}
  \item{\code{temperature}}{\link{temperatureType}.}
  \item{\code{gradient}}{\link{gradientType}.}
  \item{\code{loop}}{\link{loopType}.} \item{\code{pause}}{\link{pauseType}.}
  \item{\code{lidOpen}}{\link{lidOpenType}.}
  }
}
\keyword{datasets}

