% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{stepType}
\alias{stepType}
\title{stepType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
stepType$new(nr, description = NULL, temperature = NULL, gradient =
  NULL, loop = NULL, pause = NULL, lidOpen = NULL)
}
\arguments{
\item{nr}{\link[assertthat]{is.count}. The incremental number of the step.
First step should be nr = 1 and then increment each step by + 1.}

\item{description}{\link[assertthat]{is.string}.}

\item{temperature}{\link{temperatureType}.}

\item{gradient}{\link{gradientType}.}

\item{loop}{\link{loopType}.}

\item{pause}{\link{pauseType}.}

\item{lidOpen}{\link{lidOpenType}.}
}
\description{
Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

