% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{targetTypeType}
\alias{targetTypeType}
\title{targetTypeType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
targetTypeType
}
\description{
Can take values:
\describe{
\item{ref}{reference target}
\item{toi}{target of interest}
}
Inherits: \link{enumType}.
}
\section{Initialization}{
 \preformatted{targetTypeType$new(value)}

  @section Fields: \describe{
\item{\code{value}}{\link[checkmate]{checkString}.}
}
}
\keyword{datasets}

