% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{targetTypeType}
\alias{targetTypeType}
\title{targetTypeType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
targetTypeType$new(value)
}
\arguments{
\item{value}{\link[assertthat]{is.string}.}
}
\description{
Can take values:
\describe{
\item{ref}{reference target}
\item{toi}{target of interest}
}
\cr Inherits: \link{enumType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

