% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{temperatureType}
\alias{temperatureType}
\title{temperatureType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
temperatureType$new(temperature, ...)
}
\arguments{
\item{temperature}{\link[base]{double}. The temperature of the step in
degrees Celsius.}

\item{...}{Params of parent class \link{baseTemperatureType}.}
}
\description{
This step keeps a constant temperature on the heat block. Inherits:
\link{baseTemperatureType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

