% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{temperatureType}
\alias{temperatureType}
\title{temperatureType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
temperatureType
}
\description{
This step keeps a constant temperature on the heat block. Inherits:
\link{baseTemperatureType}.
}
\section{Initialization}{
 \preformatted{temperatureType$new(temperature, ...)}
}

\section{Fields}{
 \describe{ \item{\code{temperature}}{\link[checkmate]{checkNumber}.
  The temperature of the step in  degrees Celsius.}
\item{\code{...}}{ Params of parent class \link{baseTemperatureType}.}
}
}

\keyword{datasets}
