% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{templateQuantityType}
\alias{templateQuantityType}
\title{templateQuantityType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
templateQuantityType
}
\description{
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{templateQuantityType$new(conc, nucleotide)}
}

\section{Fields}{
 \describe{
\item{\code{conc}}{\link[base]{is.double}. Concentration of the template in nanogram}
  per microliter in the final reaction mix.
\item{\code{nucleotide}}{\link{nucleotideType}.}
  }
}
\keyword{datasets}

