% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{thermalCyclingConditionsType}
\alias{thermalCyclingConditionsType}
\title{thermalCyclingConditionsType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
thermalCyclingConditionsType
}
\description{
A cycling program for PCR or to amplify cDNA. Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \code{thermalCyclingConditionsType$new(id,
  description = NULL, documentation = NULL, lidTemperature = NULL,
  experimenter = NULL, step)}
}

\section{Fields}{
 \describe{ \item{\code{id}}{\link{idType}.}
  \item{\code{description}}{\link[assertthat]{is.string}.}
  \item{\code{documentation}}{\code{list} of \link{idReferencesType}.}
  \item{\code{lidTemperature}}{\link[base]{double}. The temperature in
  degrees Celsius of the lid during cycling. }
  \item{\code{experimenter}}{\code{list} of \link{idReferencesType}.
  Reference to the person who made or uses this protocol. }
  \item{\code{step}}{\code{list} of \link{stepType}. The steps a protocol
  runs through to amplify DNA.}
  }
}
\keyword{datasets}

