% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{xRefType}
\alias{xRefType}
\title{xRefType R6 class.}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Inherits: \link{rdmlBaseType}.
}
\section{Initialization}{
 \preformatted{xRefType$new(name = NULL, id = NULL)}

  @section Fields: \describe{
  \item{\code{name}}{\link[checkmate]{checkString}. Reference to an external
  database, } for example "GenBank".
  \item{\code{id}}{\link[checkmate]{checkString}. The ID of the entry within
  the external database, for example "AJ832138".}
  }
}

\section{Super class}{
\code{\link[RDML:rdmlBaseType]{RDML::rdmlBaseType}} -> \code{xRefType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-xRefType-new}{\code{xRefType$new()}}
\item \href{#method-xRefType-clone}{\code{xRefType$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id=".asXMLnodes"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-.asXMLnodes'><code>RDML::rdmlBaseType$.asXMLnodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="copy"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-copy'><code>RDML::rdmlBaseType$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RDML" data-topic="rdmlBaseType" data-id="print"><a href='../../RDML/html/rdmlBaseType.html#method-rdmlBaseType-print'><code>RDML::rdmlBaseType$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xRefType-new"></a>}}
\if{latex}{\out{\hypertarget{method-xRefType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xRefType$new(name = NULL, id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xRefType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-xRefType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xRefType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
