% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDML.types.R
\docType{class}
\name{xRefType}
\alias{xRefType}
\title{xRefType R6 class.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
xRefType$new(name = NULL, id = NULL)
}
\arguments{
\item{name}{\link[assertthat]{is.string}. Reference to an external database,
for example "GenBank".}

\item{id}{\link[assertthat]{is.string}. The ID of the entry within the
  external database, for example "AJ832138".}
}
\description{
\cr Inherits: \link{rdmlBaseType}.
}
\section{Fields}{
 Names, types and description of the fields are equal to the
  class arguments.
}
\keyword{datasets}

